requirements.txt
pyyaml>=6.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
python-dateutil>=2.8.0
